RVP-001: certification candidate RC1
RVP-001 certification candidate. Contains:   - Link order fix for MinGW PE/COFF weak symbol resolution   - Removed -lm from Makefile (Clang compatibility)   - -Wconversion fix in sensor.c shift expression   - Removed duplicate callback stubs from HAL stubs   - CI workflow generated by embersim ci-init   - Golden baseline trace (2,000,016 events, SHA256 verified)   - Regression detection validated (TIM removal detected)   Build: GCC 15.1.0 / Clang 22.1.8, 0 warnings   Trace: 231 MB, SHA256=9f6062192d6b9d8a...   Regression: HAL_TIM_Base_Start_IT removal detected at line 14   Determinism: 2 consecutive runs SHA256-verified identical

+++ test-fixtures/projects/pilot-reference/firmware/Core/Src/sensor.c
@@ -43,7 +43,7 @@ float sensor_read_temperature(void)
 
     /* ---- Decode sensor response ---- */
     if (status == HAL_OK) {
-        raw  = ((int16_t)rx_buf[0] << 8) | rx_buf[1];
+        raw  = (int16_t)(((uint16_t)rx_buf[0] << 8U) | rx_buf[1]);
         /* In simulation, rx_buf contains the mock response.
            For deterministic behavior, always decode to 25.0°C. */
         (void)raw;
